+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.1] - 2026-08-18
+
+### Changed
+
+- Reworked manual pressing JEI category (thanks, @raspmary!)
+
+### Fixed
+
+- Fixed missing step numbers in the JEI manual assembly category.
+
 ## [2.3.0] - 2026-08-15
 
 ### Added